+++ sdk/keyvault/Azure.Security.KeyVault.Keys/tests/CryptographyModelFactoryTests.cs
@@ -30,8 +30,7 @@ namespace Azure.Security.KeyVault.Keys.Tests
         public void DecryptOptionsAll()
         {
             byte[] buffer = new byte[] { 0, 1, 2, 3 };
-            DecryptParameters options = CryptographyModelFactory.DecryptParameters(EncryptionAlgorithm.A128Cbc, buffer, buffer, buffer);
-            options.AdditionalAuthenticatedData = buffer;
+            DecryptParameters options = CryptographyModelFactory.DecryptParameters(EncryptionAlgorithm.A128Cbc, buffer, buffer, buffer, buffer);
 
             Assert.AreEqual(EncryptionAlgorithm.A128Cbc, options.Algorithm);
             CollectionAssert.AreEqual(buffer, options.Ciphertext);
@@ -60,8 +59,7 @@ namespace Azure.Security.KeyVault.Keys.Tests
         public void EncryptOptionsAll()
         {
             byte[] buffer = new byte[] { 0, 1, 2, 3 };
-            EncryptParameters options = CryptographyModelFactory.EncryptParameters(EncryptionAlgorithm.A128Cbc, buffer, buffer);
-            options.AdditionalAuthenticatedData = buffer;
+            EncryptParameters options = CryptographyModelFactory.EncryptParameters(EncryptionAlgorithm.A128Cbc, buffer, buffer, buffer);
 
             Assert.AreEqual(EncryptionAlgorithm.A128Cbc, options.Algorithm);
             CollectionAssert.AreEqual(buffer, options.Plaintext);